+++ src/main/java/org/indigo/cloudproviderruleengine/RankedCloudProvider.java
@@ -3,9 +3,9 @@ package org.indigo.cloudproviderruleengine;
 public class RankedCloudProvider {
 	private int     id;
 	private String  name;
-	private double  rank;
+	private float   rank;
 	
-	public RankedCloudProvider( int id, String name, double rank) {
+	public RankedCloudProvider( int id, String name, float rank) {
 		this.id   = id;
 		this.name = name;
 		this.rank = rank;
